pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anyscribecli"
version = "0.7.4.3"
description = "CLI tool to download, transcribe, and convert video/audio to structured markdown"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Rish Madaan"},
]
keywords = ["transcription", "whisper", "youtube", "markdown", "obsidian", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "typer[all]>=0.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "yt-dlp>=2024.0",
    "rich>=13.0",
    "beaupy>=3.0",
    "instaloader>=4.10",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
]

[project.urls]
Repository = "https://github.com/rishmadaan/anyscribecli"
Issues = "https://github.com/rishmadaan/anyscribecli/issues"
Documentation = "https://github.com/rishmadaan/anyscribecli/tree/main/docs"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]
local = [
    "faster-whisper>=1.0",
]
mcp = [
    "mcp>=1.0",
]

[project.scripts]
scribe = "anyscribecli.cli.main:app"
ascli = "anyscribecli.cli.main:app"
scribe-mcp = "anyscribecli.mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/anyscribecli"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
